+++ zed_components/src/tools/src/sl_tools.cpp
@@ -512,27 +512,6 @@ std::vector<std::vector<float>> parseStringVector(
   return result;
 }
 
-SmartMean::SmartMean(int winSize)
-{
-  mValCount = 0;
-
-  mMeanCorr = 0.0;
-  mMean = 0.0;
-  mWinSize = winSize;
-
-  mGamma = (static_cast<double>(mWinSize) - 1.) / static_cast<double>(mWinSize);
-}
-
-double SmartMean::addValue(double val)
-{
-  mValCount++;
-
-  mMeanCorr = mGamma * mMeanCorr + (1. - mGamma) * val;
-  mMean = mMeanCorr / (1. - pow(mGamma, mValCount));
-
-  return mMean;
-}
-
 bool isZED2OrZED2i(sl::MODEL camModel)
 {
   if (camModel == sl::MODEL::ZED2) {